add integration tests

+++ src/main/java/ru/yumeno/nir/service/impl/AddressServiceImpl.java
@@ -3,8 +3,13 @@ package ru.yumeno.nir.service.impl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.yumeno.nir.entity.Address;
+import ru.yumeno.nir.entity.News;
+import ru.yumeno.nir.entity.Subscription;
+import ru.yumeno.nir.exception_handler.exceptions.DeletionFailedException;
 import ru.yumeno.nir.exception_handler.exceptions.ResourceNotFoundException;
 import ru.yumeno.nir.repository.AddressRepository;
+import ru.yumeno.nir.repository.NewsRepository;
+import ru.yumeno.nir.repository.SubscriptionRepository;
 import ru.yumeno.nir.service.AddressService;
 
 import java.util.List;
@@ -13,10 +18,14 @@ import java.util.Optional;
 @Service
 public class AddressServiceImpl implements AddressService {
     private final AddressRepository addressRepository;
+    private final NewsRepository newsRepository;
+    private final SubscriptionRepository subscriptionRepository;
 
     @Autowired
-    public AddressServiceImpl(AddressRepository addressRepository) {
+    public AddressServiceImpl(AddressRepository addressRepository, NewsRepository newsRepository, SubscriptionRepository subscriptionRepository) {
         this.addressRepository = addressRepository;
+        this.newsRepository = newsRepository;
+        this.subscriptionRepository = subscriptionRepository;
     }
 
     @Override
@@ -43,6 +52,15 @@ public class AddressServiceImpl implements AddressService {
 
     @Override
     public void deleteAddress(int id) {
-        addressRepository.deleteById(id);
+        Optional<Address> optional = addressRepository.findById(id);
+        List<News> news = newsRepository.findAllByAddresses(optional.
+                orElseThrow(() -> new ResourceNotFoundException("Address not exist with id : " + id)));
+        List<Subscription> subscriptions = subscriptionRepository.findAllByAddress(optional.
+                orElseThrow(() -> new ResourceNotFoundException("Address not exist with id : " + id)));
+        if (news.isEmpty() && subscriptions.isEmpty()) {
+            addressRepository.deleteById(id);
+        } else {
+            throw new DeletionFailedException("Address cannot be deleted with id : " + id);
+        }
     }
 }
