+++ src/main/java/ru/yumeno/nir/service/impl/DistrictServiceImpl.java
@@ -2,8 +2,11 @@ package ru.yumeno.nir.service.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.yumeno.nir.entity.Address;
 import ru.yumeno.nir.entity.District;
+import ru.yumeno.nir.exception_handler.exceptions.DeletionFailedException;
 import ru.yumeno.nir.exception_handler.exceptions.ResourceNotFoundException;
+import ru.yumeno.nir.repository.AddressRepository;
 import ru.yumeno.nir.repository.DistrictRepository;
 import ru.yumeno.nir.service.DistrictService;
 
@@ -13,10 +16,12 @@ import java.util.Optional;
 @Service
 public class DistrictServiceImpl implements DistrictService {
     private final DistrictRepository districtRepository;
+    private final AddressRepository addressRepository;
 
     @Autowired
-    public DistrictServiceImpl(DistrictRepository districtRepository) {
+    public DistrictServiceImpl(DistrictRepository districtRepository, AddressRepository addressRepository) {
         this.districtRepository = districtRepository;
+        this.addressRepository = addressRepository;
     }
 
     @Override
@@ -43,6 +48,13 @@ public class DistrictServiceImpl implements DistrictService {
 
     @Override
     public void deleteDistrict(int id) {
-        districtRepository.deleteById(id);
+        Optional<District> optional = districtRepository.findById(id);
+        List<Address> addresses = addressRepository.findAllByDistrict(optional.
+                orElseThrow(() -> new ResourceNotFoundException("District not exist with id : " + id)));
+        if (addresses.isEmpty()) {
+            districtRepository.deleteById(id);
+        } else {
+            throw new DeletionFailedException("District cannot be deleted with id : " + id);
+        }
     }
 }
